Fixed some warnings and names.

+++ x3daudio1_7/wave/WaveFile.h
@@ -9,7 +9,7 @@ class WaveFile
 public:
 	WaveFile(const std::wstring & file_name, unsigned int num_channels, unsigned int samples_per_second, int bits_per_sample);
 	~WaveFile();
-	void AppendData(const uint8_t * buffer, size_t length);
+	void AppendData(const uint8_t * buffer, uint32_t length);
 
 private:
 	std::ofstream m_stream;
